changed floats to integers

+++ PositionComponent.h
@@ -7,8 +7,8 @@ public:
 	PositionComponent(int a, int b) { x = a; y = b; }
 	int getXPosition() { return x; }
 	int getYPosition() { return y; }
-	void setXPosition(float a) { x = a; }
-	void setYPosition(float b) { y = b; }
+	void setXPosition(int a) { x = a; }
+	void setYPosition(int b) { y = b; }
 	TYPE getType() { return type; }
 private:
 	TYPE type = TYPE::POSITION;
